#39 Move eo3 code from app to lib
